refactor: transaction mode works and propagates
operations on counters moved to new file, dipatching transaction mode
functions properly with custom 'UNDO' command
